+++ src/pages/CounterPage.js
@@ -1,21 +1,14 @@
-import { useState, useEffect } from "react";
 import Button from "../components/Button";
+import useCounter from "../hooks/use-counter";
+
 
 function CounterPage({ intialCount }) {
-  const [count, setCount] = useState(intialCount);
-  
-  useEffect(() => {
-    console.log(count);
-  }, [count]);
-
-  const handleClick = () => {
-    setCount(count + 1);
-  };
+  const [count, updateCount] = useCounter(intialCount) ;
 
   return (
     <div>
       <h1>Counter : {count} </h1>
-      <Button onClick={handleClick}> Increment</Button>
+      <Button onClick={updateCount}> Increment</Button>
     </div>
   );
 }
